Lots and lots of fixes - upgraded to gst-plugins-bad 0.10.17, gst-plugins-ugly 0.10.13, gst-plugins-gl 0.10.1
git-svn-id: http://ossbuild.googlecode.com/svn/trunk@365 72a73a14-1cf0-11de-a027-6f5b8e30e268

+++ Main/GStreamer/Source/gst-plugins-good/sys/osxaudio/gstosxringbuffer.c
@@ -43,6 +43,7 @@
  */
 
 #include <CoreAudio/CoreAudio.h>
+#include <CoreServices/CoreServices.h>
 #include <gst/gst.h>
 #include <gst/audio/multichannel.h>
 #include "gstosxringbuffer.h"
@@ -190,7 +191,8 @@ gst_osx_ring_buffer_create_audio_unit (GstOsxRingBuffer * osxbuf,
 
     if (status) {
       CloseComponent (unit);
-      GST_WARNING_OBJECT (osxbuf, "Failed to enable input: %lx", status);
+      GST_WARNING_OBJECT (osxbuf, "Failed to enable input: %lx",
+          (gulong) status);
       return NULL;
     }
 
@@ -201,7 +203,8 @@ gst_osx_ring_buffer_create_audio_unit (GstOsxRingBuffer * osxbuf,
 
     if (status) {
       CloseComponent (unit);
-      GST_WARNING_OBJECT (osxbuf, "Failed to disable output: %lx", status);
+      GST_WARNING_OBJECT (osxbuf, "Failed to disable output: %lx",
+          (gulong) status);
       return NULL;
     }
   }
@@ -213,7 +216,7 @@ gst_osx_ring_buffer_create_audio_unit (GstOsxRingBuffer * osxbuf,
 
   if (status) {
     CloseComponent (unit);
-    GST_WARNING_OBJECT (osxbuf, "Failed to set device: %lx", status);
+    GST_WARNING_OBJECT (osxbuf, "Failed to set device: %lx", (gulong) status);
     return NULL;
   }
 
@@ -251,7 +254,7 @@ gst_osx_ring_buffer_open_device (GstRingBuffer * buf)
       CloseComponent (osxbuf->audiounit);
       osxbuf->audiounit = NULL;
       GST_WARNING_OBJECT (osxbuf, "Unable to obtain device properties: %lx",
-          status);
+          (gulong) status);
       return FALSE;
     }
 
@@ -399,7 +402,8 @@ gst_osx_ring_buffer_acquire (GstRingBuffer * buf, GstRingBufferSpec * spec)
       kAudioUnitProperty_StreamFormat, scope, element, &format, propertySize);
 
   if (status) {
-    GST_WARNING_OBJECT (osxbuf, "Failed to set audio description: %lx", status);
+    GST_WARNING_OBJECT (osxbuf, "Failed to set audio description: %lx",
+        (gulong) status);
     goto done;
   }
 
@@ -408,7 +412,7 @@ gst_osx_ring_buffer_acquire (GstRingBuffer * buf, GstRingBufferSpec * spec)
       scope, element, layout, layoutSize);
   if (status) {
     GST_WARNING_OBJECT (osxbuf, "Failed to set output channel layout: %lx",
-        status);
+        (gulong) status);
     goto done;
   }
 
@@ -424,7 +428,8 @@ gst_osx_ring_buffer_acquire (GstRingBuffer * buf, GstRingBufferSpec * spec)
         &frameSize, &propertySize);
 
     if (status) {
-      GST_WARNING_OBJECT (osxbuf, "Failed to get frame size: %lx", status);
+      GST_WARNING_OBJECT (osxbuf, "Failed to get frame size: %lx",
+          (gulong) status);
       goto done;
     }
 
@@ -546,7 +551,7 @@ gst_osx_ring_buffer_start (GstRingBuffer * buf)
   osxbuf = GST_OSX_RING_BUFFER (buf);
 
   GST_DEBUG ("osx ring buffer start ioproc: 0x%p device_id %lu",
-      osxbuf->element->io_proc, osxbuf->device_id);
+      osxbuf->element->io_proc, (gulong) osxbuf->device_id);
   if (!osxbuf->io_proc_active) {
     callback_type = osxbuf->is_src ?
         kAudioOutputUnitProperty_SetInputCallback :
@@ -591,7 +596,7 @@ gst_osx_ring_buffer_pause (GstRingBuffer * buf)
   GstOsxRingBuffer *osxbuf = GST_OSX_RING_BUFFER (buf);
 
   GST_DEBUG ("osx ring buffer pause ioproc: 0x%p device_id %lu",
-      osxbuf->element->io_proc, osxbuf->device_id);
+      osxbuf->element->io_proc, (gulong) osxbuf->device_id);
   if (osxbuf->io_proc_active) {
     /* CoreAudio isn't threadsafe enough to do this here; we must deactivate
      * the render callback elsewhere. See:
@@ -612,7 +617,7 @@ gst_osx_ring_buffer_stop (GstRingBuffer * buf)
   osxbuf = GST_OSX_RING_BUFFER (buf);
 
   GST_DEBUG ("osx ring buffer stop ioproc: 0x%p device_id %lu",
-      osxbuf->element->io_proc, osxbuf->device_id);
+      osxbuf->element->io_proc, (gulong) osxbuf->device_id);
 
   status = AudioOutputUnitStop (osxbuf->audiounit);
   if (status)
